assignment 7

+++ server/models/note.js
@@ -16,40 +16,58 @@ createTable();
 async function getAllNotes() {
    const sql = `SELECT * FROM notes;`;
    let notes = await con.query(sql);
-   console.log(notes)
+   return await con.query(sql);
+
 }
 
-getAllNotes();
-updateNote(1,'hello doremon');
+// getAllNotes();
+// updateNote(1,'hello doremon');
+
+async function CreateNote(note){
+  let cNote = await getNote(note.noteID);
+  if(cNote.length > 0) throw Error("noteId already in use");
+
+  const sql= `INSERT INTO notes (userID,noteID,emailId, noteContent)
+    VALUES (${note.userID}, "${note.noteID}","${note.emailId}","${note.noteContent}");
+  `
+  await con.query(sql);
+  return await getNote(note.noteID);
+}
 
 
-async function getNote(noteId) {
+
+async function getNote(noteID) {
   let sql = `
     SELECT * FROM notes
-      WHERE noteId = "${noteId}"
+      WHERE noteID = "${noteID}"
   `;
 
   return await con.query(sql);
 }
 
-async function deleteNote(noteID) {
+async function deleteNote(note) {
+
+  let cNote = await getNote(note.noteID);
+   if(!cNote[0]) throw Error("Note not found");
+
   let sql = `
     DELETE FROM notes
-      WHERE noteID = "${noteID}"
+      WHERE noteID = "${note.noteID}"
   `;
 
   return await con.query(sql);
 }
 
-deleteNote(1);
+//deleteNote(1);
 
-async function updateNote(noteID, noteContent) {
+async function updateNote(note) {
+  let cNote = await getNote(note.noteID);
+   if(!cNote[0]) throw Error("Note not found");
+   
   let sql = `
-    UPDATE notes SET noteContent = "${noteContent}"
-      WHERE noteID = "${noteID}"
+    UPDATE notes SET noteContent = "${note.noteContent}"
+      WHERE noteID = "${note.noteID}"
   `;
-
-
 
   return await con.query(sql);
 }
@@ -57,4 +75,4 @@ async function updateNote(noteID, noteContent) {
 
 
 
-module.exports = { getAllNotes, updateNote, deleteNote, getNote };
+module.exports = { getAllNotes, updateNote, deleteNote, getNote, CreateNote };
